+++ th2_data_services/data_source/lwdp/resolver.py
@@ -79,8 +79,16 @@ class MessageFieldResolver(resolver_core.MessageFieldResolver):
         raise NotImplementedError
 
     @staticmethod
-    def get_sequence(message) -> str:  # <book>:<alias>:<direction>:<timestamp>:<sequence>.<\d>.<\d>
-        return message[message_struct.MESSAGE_ID].split(":")[4].split(".")[0]
+    def get_sequence(
+        message,
+    ) -> str:  # <book>:<group>:<alias>:<direction>:<timestamp>:<sequence>.<\d>.<\d>
+        return message[message_struct.MESSAGE_ID].split(":")[5].split(".")[0]
+
+    @staticmethod
+    def get_group(
+        message,
+    ) -> str:  # <book>:<group>:<alias>:<direction>:<timestamp>:<sequence>.<\d>.<\d>
+        return message[message_struct.MESSAGE_ID].split(":")[1]
 
     @staticmethod
     def get_timestamp(message) -> Dict[str, int]:
@@ -166,7 +174,13 @@ class ExpandedMessageFieldResolver(resolver_core.ExpandedMessageFieldResolver):
 
     @staticmethod
     def get_sequence(message) -> str:  # <book>:<alias>:<direction>:<timestamp>:<sequence>.<\d>.<\d>
-        return message[message_struct.MESSAGE_ID].split(":")[4].split(".")[0]
+        return message[message_struct.MESSAGE_ID].split(":")[5].split(".")[0]
+
+    @staticmethod
+    def get_group(
+        message,
+    ) -> str:  # <book>:<group>:<alias>:<direction>:<timestamp>:<sequence>.<\d>.<\d>
+        return message[message_struct.MESSAGE_ID].split(":")[1]
 
     @staticmethod
     def get_timestamp(message) -> Dict[str, int]: